requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Pyrebase4==4.5.0
pika==1.2.0